requirements.txt
pyg-lib
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
-f https://data.pyg.org/whl/torch-1.13.0+cu116.html
pytorch-lightning
matplotlib
momo-opt
feather-format